+++ kernel/LibC/vga.h
@@ -27,5 +27,5 @@ void vga_putchar(char c);
 void vga_print(const char* s);
 void vga_input(uint32_t tecla);
 void vga_pulaLinha(uint32_t tecla);
-void vga_uint32(uint8_t tecla);
+void vga_uint32(uint32_t tecla);
 #endif
